XRDDEV-766 Split token into a new component
